Only kill the simulator when it's running

+++ tests/integration-tests/helix-payload/osx-helix-payload.sh
@@ -41,7 +41,10 @@ export XHARNESS_LOG_WITH_TIMESTAMPS=true
 # Restart the simulator to make sure it is tied to the right user session
 xcode_path=`xcode-select -p`
 pid=`ps aux | grep $xcode_path/Applications/Simulator.app | grep -v grep | tr -s ' ' | cut -d ' ' -f 2`
-sudo kill $pid
+if [ -z "$pid" ]; then
+    sudo kill $pid
+fi
+
 open -a $xcode_path/Applications/Simulator.app
 
 dotnet tool restore --no-cache
